% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepass.R
\name{op_get_item}
\alias{op_get_item}
\title{Retrieve a username and password}
\usage{
op_get_item(ops, name, fields = "username,password")
}
\arguments{
\item{ops}{1Password Session object, created from setup_op() or unlock_op()}

\item{name}{Specify an item by name or uuid. UUID is preferred for consistency.}

\item{fields}{Optional. List of fields desired. NULLable to get all contents.}
}
\value{
A character vector
}
\description{
Retrieve a username and password
}
