% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepass.R
\name{op_list_items}
\alias{op_list_items}
\title{List items}
\usage{
op_list_items(ops, vault = NULL)
}
\arguments{
\item{ops}{1Password Session object, created from setup_op() or unlock_op()}

\item{vault}{Optional. Can specify a vault by name or uuid to see passwords
within a specific vault}
}
\value{
A data frame of password items with titles, usernames and uuids
}
\description{
List items
}
