% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepass.R
\name{op_list_vaults}
\alias{op_list_vaults}
\title{List vaults}
\usage{
op_list_vaults(ops)
}
\arguments{
\item{ops}{1Password Session object, created from setup_op() or unlock_op()}
}
\value{
A data frame containing the vaults available in the session
}
\description{
List vaults
}
