% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepass.R
\name{unlock_op}
\alias{unlock_op}
\title{Unlock vault}
\usage{
unlock_op(
  domain,
  masterpassword = rstudioapi::askForPassword("Please enter your Master Password")
)
}
\arguments{
\item{domain}{Full domain URL of your vault}

\item{masterpassword}{Used to unlock the vault}
}
\value{
1Password session object, to be used for interacting with vault
}
\description{
Saves a token as an environment variable for the session as OP_SESSION_(subdomain)
}
